// frontend/src/shared/api/social/model/reportRequestReason.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * defShows Social API
 * OpenAPI spec version: 0.1.0
 */

export type ReportRequestReason = typeof ReportRequestReason[keyof typeof ReportRequestReason];


export const ReportRequestReason = {
  spam: 'spam',
  harassment: 'harassment',
  inappropriate: 'inappropriate',
  other: 'other',
} as const;